Fixed a few issues
-Prevent client crash when Dear Imgui has 0 draws
-Fixed 32bits Client rendering issues
-Added Dear ImGui 1.87 compatibility test

+++ Code/Client/Private/NetImgui_CmdPackets.inl
@@ -51,8 +51,8 @@ bool CmdInput::IsKeyDown(eVirtualKeys vkKey)const
 
 void CmdInput::SetKeyDown(eVirtualKeys vkKey, bool isDown)
 {
-	const size_t keyEntryIndex	= static_cast<uint64_t>(vkKey) / 64;
-	const uint64_t keyBitMask	= static_cast<uint64_t>(1) << static_cast<uint64_t>(vkKey) % 64;	
+	const uint64_t keyEntryIndex	= static_cast<uint64_t>(vkKey) / 64;
+	const uint64_t keyBitMask		= static_cast<uint64_t>(1) << static_cast<uint64_t>(vkKey) % 64;	
 	mKeysDownMask[keyEntryIndex]= isDown ?	mKeysDownMask[keyEntryIndex] | keyBitMask : 
 											mKeysDownMask[keyEntryIndex] & ~keyBitMask;
 }
